+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add detailed reason to the warning panel when an avatar is not uploadable for Quest on Android build target.
 - Add zip file for VCC/VPM.
+- Add Animator Controller conversion also for Merge Animator of Modular Avatar.
 
 ### Changed
 - Change updater info to prevent opening Booth link when using as a VPM package.